+++ crispy_print/public/js/components/FieldsPane.vue
@@ -3,7 +3,7 @@
 		<div class="section-head fields-pane__header">
 			<div class="section-head-content fields-pane__header-row">
 				<h3 class="section-title fields-pane__title">
-					{{ __("Fields") }}
+					{{ __("Fields", null, "Crispy Print UI") }}
 				</h3>
 				<div class="fields-pane__spacer"></div>
 				<div class="fields-pane__help">